Add GDI font provider
The get_fallback doesn't always work. It is possible to use GDI+ to find a right fallback, but for unknown reason, it only works on Windows 8 or more.

The code should be properly refactored. This is just a proof of concept.
Also, this new provider is currently inly supported with meson, not autotools.