Modify GetFallbackFunc to return ASS_FontInfo + add bold/italic parameter
The big problem with this solution is that each time we request a fallback_font, the font (even if we have previously requested it) gets added to the provider database.

As a result, we could end up with the exact same font duplicated 100 times in the database.

Currently, I don't see to properly implement this.
Note that we need to return ASS_FontInfo* because, for the android provider, we can't get the font family.